clean up repository, fix code first ef db generation

+++ WebAPI/WebAPI/Data/DbInitializer.cs
@@ -1,9 +1,16 @@
-﻿namespace WebAPI.Data
+﻿using System.Linq;
+
+namespace WebAPI.Data
 {
     public static class DbInitializer
     {
         public static void Seed(WorkContext context)
         {
+            if (context.WorkInquiries.Any())
+            {
+                return;
+            }
+
             context.Database.EnsureDeleted();
             context.Database.EnsureCreated();
             context.SaveChanges();
